Move component classes out
